requirements.txt
boto3==1.34.51
python-dotenv==1.0.1
pyspark==3.5.0
jupyter==1.0.0